+++ apps/web/src/features/user/hooks/use-preferred-time-format.ts
@@ -1,8 +1,8 @@
 "use client";
 
 import { useUser } from "@/components/user-provider";
+import { getLocaleDefaults } from "@/lib/datetime/locales";
 import { useLocale } from "@/lib/locale/client";
-import { getLocaleDefaults } from "@/lib/localization/locales";
 
 export function usePreferredTimeFormat() {
   const { user } = useUser();
